web: add sanity check for service worker message order
I couldn't find anything that clearly stated calls to postMessage()
are guaranteed to be received in the same order, and I don't
understand the JS event loop well enough to work it out myself.

Instead of risking silently corrupting files, this commit add a
sanity check to make sure the chunks are being written at the
expected offset. Better paranoid and safe than clever and sorry.