cmd: switch websocket package
The reason for the switch is that the new package can be compiled
to WASM.

This opens up moving more of the handshake from js to wasm. Alas,
the wasm binary with webrtc & websockets packages linked in is in
excess of 8MB, ~3MB compressed.

I'll keep this websocket package to continue experimenting with
the idea but stop short of switching. 800KB of wasm is already
a lot.