verbose logging
This adds a -verbose flag to the command line client, and prints
a lot more to the web console.

This currently includes all candidates, offers and answers. To make
it easier for users to copy and paste these logs we should filter
out ip addresses or just print out a summary.