server: add turn auth support
This adds options to the signal server to generate short-term
credentials for TURN servers given a shared secret.

https://tools.ietf.org/html/draft-uberti-behave-turn-rest-00

The scheme, tested with the coturn server, is as follows:

	username: (expiry time in unix seconds) + ":" + user_id
	password: base64(sha1hmac(secret, username))

user_id does not seem to matter, except maybe for quota allocations.

As a result of the signalling server now telling the clients what
ICE servers to use, and WebRTC's PeerConnection API design, the
PeerConnection initialisation code now has to be intertwined with
the signalling handshake.

This makes this package's high-level API a real mess. I will revisit
this once I've thought about it more. At least TURN works now.

N.B. there seems to be a SetConfiguration method on PeerConnection
objects. This would make our API much nicer but it seems ineffective
for resetting ICE servers on at least Firefox and Pion.