switch away from pgp word list
some of the pgp words seem to have questionable appropriateness.
this changes the list used to the soundex filtered eff list devised
by @nickzoic (https://nick.zoic.org/art/shorter-words-list/).

as an experiment, this also changes the <number>-<words> format in
favour of a list of words beginning with the slot number encoded
as a uvarint.

this also introduces some refactoring which can pave the way to
have multiple encodings in the future.

see #8 for further discussion.